requirements.txt
# Основні залежності
yt-dlp>=2023.12.30
openai-whisper>=20231117
ffmpeg-python>=0.2.0

# Обробка аудіо та відео
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.22.0

# Обробка тексту
tiktoken>=0.5.1
regex>=2023.0.0
transformers>=4.30.0
sentencepiece>=0.1.99

# Взаємодія з API
requests>=2.31.0
aiohttp>=3.8.0

# Утиліти
tqdm>=4.65.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Опціональні залежності для розробки
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0 
crewai>=0.1.1
